+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on the [KeepAChangeLog] project.
 
 ### Changed
 - [#827] Added support for python 3.11
+- [#830], [#831] Allow null and empty values in UserInfo responses, but filter them out. 
 
 ### Fixed
 - [#826], [#829] Fixed RP-Initiated Logout To Accept id_token_hint
@@ -21,6 +22,8 @@ The format is based on the [KeepAChangeLog] project.
 [#827]: https://github.com/OpenIDC/pyoidc/issues/827
 [#826]: https://github.com/OpenIDC/pyoidc/issues/826
 [#829]: https://github.com/OpenIDC/pyoidc/pull/829
+[#830]: https://github.com/OpenIDC/pyoidc/issues/830
+[#831]: https://github.com/OpenIDC/pyoidc/pull/831
 
 ## 1.4.0 [2022-05-23]